Create tutorial route
